Rename Person collection to match application naming conventions
