+++ view/app/imports/api/person/index.js
@@ -1,0 +1,2 @@
+import './PersonCollection.js';
+import './PersonCollectionMethods.js';
