--- view/app/imports/api/persons/index.js
@@ -1,2 +1,0 @@
-import './persons.js';
-import './personsMethods.js';
