+++ view/app/imports/startup/both/index.js
@@ -4,7 +4,7 @@ import '../../utils/schemas.js';
 
 import '../../api/measurement';
 import '../../api/simulatedEvents';
-import '../../api/persons';
+import '../../api/person';
 import '../../api/boxEvent';
 import '../../api/eventData';
 import '../../api/eventMetaData';
